refact:dismantle Banner, recompose to NewShared/-custom, part of Wrapper
