--- app/Within/pageFocus/Main/MainBanner/styles.module.css
@@ -1,152 +1,0 @@
-.comMainBanner {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: 100%;
-  position: relative;
-  box-sizing: border-box;
-}
-
-.boxRow {
-  width: 100%;
-  margin: 5vw 0 7.5vw;
-}
-
-.boxRowBelong {
-  align-self: center;
-  position: relative;
-  box-sizing: border-box;
-}
-
-.boxRowTodayNode {
-  position: relative;
-  box-sizing: border-box;
-}
-
-.boxRowBroads {
-  position: relative;
-  box-sizing: border-box;
-}
-
-.boxRowFirst {
-  display: flex;
-  justify-content: space-between;
-  position: relative;
-  box-sizing: border-box;
-  margin: 5vw 0 7.5vw;
-}
-
-.boxRowTitle {
-  width: 100%;
-  box-sizing: border-box;
-}
-
-.boxUnits {
-  display: flex;
-  justify-content: space-between;
-  flex-wrap: wrap;
-  width: 100%; /*controled by parent*/
-  position: relative;
-  box-sizing: border-box;
-}
-
-.boxTitle {
-  display: flex;
-  justify-content: left;
-  width: 100%;
-  box-sizing: border-box;
-  margin: 5vw 0 0vw;
-}
-
-.rowBroads {
-
-}
-
-.boxTitleFirst {
-  width: 20vw;
-  height: 30vw; /*follow current heightNine*/
-  position: relative;
-}
-
-.comTodayNode {
-  display: flex;
-  justify-content: space-between;
-  width: 100%;
-}
-
-.boxTodayNails {
-
-}
-
-.boxWikiIntro {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-}
-
-.boxIntroHead {
-  width: 100%;
-  min-height: 7.5vw;
-  position: relative;
-  margin-bottom: 1rem;
-}
-
-.boxWikiTitle {
-  display: inline-flex;
-  flex-direction: column;
-  max-width: 60%;
-}
-
-.pWikiIntro {
-  width: 35vw;
-  box-sizing: border-box;
-  margin: 0;  /*<p> has its default, so overlap it*/
-}
-
-.linkMore {
-
-}
-
-.linkWiki {
-  float: right;
-}
-
-.spanTodayDescript {
-   color: rgb(110, 110, 110); /* keep it related to fontType in BannerBelong*/
-}
-
-.spanTitle {
-  display: inline-block;
-  padding: 0 0 1.76rem 0;
-}
-
-.spanTitleFirst {
-  display: block;
-  overflow: visible;
-}
-
-.fontTitle {
-  font-size: 2.2rem;
-  font-weight: 500;
-  letter-spacing: 0.064rem;
-}
-
-.fontTitleFirst {
-  font-family: 'gill-sans-nova', 'Noto Sans TC','Lato', sans-serif;
-  font-weight: 700;
-}
-
-.fontDescript { /* keep it related to fontType in BannerBelong*/
-  font-size: 1.37rem;
-  letter-spacing: 0.07rem;
-  line-height: 2rem;
-}
-
-.fontWikiIntro {
-  font-size: 1.45rem;
-  line-height: 1.84rem;
-}
-
-.fontLink {
-  font-size: 1.3rem;
-}
