+++ app/Within/pageFocus/Main/NewSharedCustom/NewSharedCustom.jsx
@@ -1,0 +1,71 @@
+import React from 'react';
+import {
+  Route,
+  Link,
+  withRouter,
+} from 'react-router-dom';
+import {connect} from "react-redux";
+import classnames from 'classnames';
+import styles from "./styles.module.css";
+import {
+  axios_Units,
+  nailChart,
+} from '../utils.js';
+import {
+  handleNounsList,
+  handleUsersList
+} from "../../../redux/actions/general.js";
+import {
+  cancelErr,
+  uncertainErr
+} from '../../../utils/errHandlers.js';
+
+class NewSharedCustom extends React.Component {
+  constructor(props){
+    super(props);
+    this.state = {
+      axios: false,
+      unitsBasic: {},
+      marksBasic: {}
+    };
+    this.axiosSource = axios.CancelToken.source();
+
+    this.style={
+
+    }
+  }
+
+
+  componentDidUpdate(prevProps, prevState, snapshot){
+
+  }
+
+  componentDidMount() {
+
+  }
+
+  componentWillUnmount() {
+
+  }
+
+
+  render(){
+
+    return(
+
+    )
+  }
+}
+
+const mapStateToProps = (state)=>{
+
+}
+
+const mapDispatchToProps = (dispatch) => {
+
+}
+
+export default withRouter(connect(
+  mapStateToProps,
+  mapDispatchToProps
+)(NewSharedCustom));
